backend/requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
asyncpg==0.27.0
certifi==2022.9.24
ciso8601==2.2.0
dnspython==2.2.1
email-validator==1.3.0
Faker==15.1.0
greenlet==1.1.3.post0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
iso8601==1.1.0
orjson==3.8.0
pydantic==1.10.2
pydantic-factories==1.9.0
pydantic-openapi-schema==1.3.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.21.0
starlite==1.35.0
starlite-multipart==1.2.0
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvloop==0.17.0
xeger==0.3.5
